Guarantee availability of Tidal, SuperDirt to Vim Plugin
Switches to a newer branch of the `vim-tidal` plugin that supports
auto-starting SuperDirt in a terminal next to the GHCi terminal. This is
disabled by default to match the original behaviour, but can be enabled
with `let g:tidal_superdirt_enable = 1`.

Also updates the `vim-tidal` defaults to use the full paths to the Tidal
GHCi instance, Tidal boot file and superdirt-start command within the
nix store. This allows for `vim-tidal` to function even outside of the
tidal devShell.